requirements.txt
Flask>=2.0.0
opencv-python>=4.5.0
tensorflow>=2.20.0
numpy>=1.21.0
Pillow>=8.0.0